# backend/docs/swagger.yaml
basePath: /
definitions:
  controller.ADregisterRequest:
    properties:
      area:
        type: string
      category:
        type: string
      description:
        type: string
      elevator:
        type: string
      floor:
        type: string
      image:
        type: string
      long:
        type: number
      lt:
        type: number
      numberOfRooms:
        type: string
      ownerid:
        type: string
      parking:
        type: string
      price:
        type: string
      store:
        type: string
      title:
        type: string
      yearOfConstruction:
        type: string
    type: object
  core.AD:
    properties:
      YearOfConstruction:
        type: integer
      area:
        type: number
      category:
        type: string
      description:
        type: string
      elevator:
        type: boolean
      floor:
        type: string
      id:
        type: integer
      long:
        type: number
      lt:
        type: number
      numberOfRooms:
        type: integer
      parking:
        type: boolean
      price:
        type: integer
      store:
        type: boolean
      title:
        type: string
      userid:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: SE Project MelkOnline backend API
  title: MelkOnline API
  version: "1.0"
paths:
  /api/v1/ads/{ad_id}:
    get:
      consumes:
      - application/json
      description: get post by id
      parameters:
      - description: post id
        in: path
        name: ad_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.AD'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: get post by id
      tags:
      - getpost
  /api/v1/ads/{ad_id}/chats:
    get:
      consumes:
      - application/json
      description: Get message from the user
      parameters:
      - description: Ad ID
        in: path
        name: ad_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: message received
          schema:
            type: string
        "400":
          description: error in reading message
          schema:
            type: string
      summary: Get message
      tags:
      - chat
    post:
      consumes:
      - application/json
      description: Send message to the user
      parameters:
      - description: Ad ID
        in: path
        name: ad_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: message sent
          schema:
            type: string
        "400":
          description: error in sending message
          schema:
            type: string
      summary: Send message
      tags:
      - chat
  /api/v1/ads/mainpage/:
    get:
      consumes:
      - application/json
      description: get ads from database to view in mainpage
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: get ads to view in mainpage
      tags:
      - mainpage
  /api/v1/ads/register:
    post:
      consumes:
      - application/json
      description: registration of ADs and giving each an ID
      parameters:
      - description: Advertisement data
        in: body
        name: ad
        required: true
        schema:
          $ref: '#/definitions/controller.ADregisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: registration of ADs
      tags:
      - ADregister
  /api/v1/ads/searchfiltering/:
    get:
      consumes:
      - application/json
      description: filter ads and get the desired ads from database
      parameters:
      - description: Category to filter
        in: body
        name: Category
        required: true
        schema:
          type: string
      - description: Max Price to filter
        in: body
        name: MaxPrice
        required: true
        schema:
          type: integer
      - description: Min Price to  filter
        in: body
        name: MinPrice
        required: true
        schema:
          type: integer
      - description: Max Area to filter
        in: body
        name: MaxArea
        required: true
        schema:
          type: integer
      - description: Min Area to filter
        in: body
        name: MinArea
        required: true
        schema:
          type: integer
      - description: Number of Rooms to filter
        in: body
        name: NumberOfRooms
        required: true
        schema:
          type: integer
      - description: Max age to filter
        in: body
        name: MaxAge
        required: true
        schema:
          type: integer
      - description: Max Floor to filter
        in: body
        name: MaxFloor
        required: true
        schema:
          type: integer
      - description: Min Floor to filter
        in: body
        name: MinFloor
        required: true
        schema:
          type: integer
      - description: availability of Elevator
        in: body
        name: Elevator
        required: true
        schema:
          type: boolean
      - description: availability of Store
        in: body
        name: Store
        required: true
        schema:
          type: boolean
      - description: availability of Parking
        in: body
        name: Parking
        required: true
        schema:
          type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Searchfiltering
      tags:
      - searchfiltering
  /api/v1/login/:
    post:
      consumes:
      - application/json
      description: get user's email and password and Login
      parameters:
      - description: user's Email
        in: body
        name: Email
        required: true
        schema:
          type: string
      - description: user's password
        in: body
        name: Password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: user Login
      tags:
      - login
  /api/v1/signup/:
    post:
      consumes:
      - application/json
      description: get user's email and password and name and signup
      parameters:
      - description: user's Email
        in: body
        name: Email
        required: true
        schema:
          type: string
      - description: user's password
        in: body
        name: Password
        required: true
        schema:
          type: string
      - description: user's password
        in: body
        name: Name
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: user signup
      tags:
      - signup
swagger: "2.0"